# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application Test",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/accounts": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Fetch all accounts",
                "description": "Returns all accounts",
                "operationId": "listAccounts",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "for search data",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Accounts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Account"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "Add a new Accounts",
                "description": "Adding new account data",
                "operationId": "addAccount",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "account_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing account parameter"
                    }
                }
            }
        },
        "/api/admin/accounts/{accountID}": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Get account by ID",
                "description": "Returns a single account data",
                "operationId": "getAccountByID",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "accountID",
                        "in": "path",
                        "description": "ID of account to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Account"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "account"
                ],
                "summary": "Edit account",
                "description": "Updating account data",
                "operationId": "editAccount",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "accountID",
                        "in": "path",
                        "description": "ID of account post to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "account_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing account parameter"
                    },
                    "404": {
                        "description": "Account not found"
                    },
                    "400": {
                        "description": "Invalid ID account"
                    }
                }
            },
            "delete": {
                "tags": [
                    "account"
                ],
                "summary": "Delete account based on ID",
                "description": "",
                "operationId": "deleteAccountById",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "accountID",
                        "in": "path",
                        "description": "ID of account to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Account": {
                "title": "Account",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "account_type": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}